backend/requirements.txt
fastapi==0.115.7
fastembed==0.5.0
langchain==0.2.2
langchain-community==0.2.3
langchain-core==0.2.43
langchain-openai==0.1.8
langchain-text-splitters==0.2.4
nltk==3.9.1
pydantic==2.10.6
python-dotenv==1.0.1
redis==5.2.1
Requests==2.32.3
uvicorn==0.34.0
chromadb==0.5.0